% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{add_title_to_yaml}
\alias{add_title_to_yaml}
\title{Add title to YAML}
\usage{
add_title_to_yaml(file)
}
\arguments{
\item{file}{Character. Path to the markdown file}
}
\description{
In the absence of a title, the file name is taken.
}
