% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{add_yaml}
\alias{add_yaml}
\title{Add YAML to file}
\usage{
add_yaml(file)
}
\arguments{
\item{file}{Character. Path to the markdown file}
}
\description{
Pre-pend YAML to file contents.
YAML contains file name as title.
}
