% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reference_index.R
\name{build_custom_reference_index}
\alias{build_custom_reference_index}
\title{Build custom reference index}
\usage{
build_custom_reference_index(dir_in, dir_out)
}
\arguments{
\item{dir_in}{Character. Path to root of Stata package directory.}

\item{dir_out}{Character. Reference foler of target documentation site.}
}
\description{
Build custom reference index
}
