% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_files}
\alias{build_files}
\title{Build Quarto files from Markdown}
\usage{
build_files(dir_in, dir_out)
}
\arguments{
\item{dir_in}{Character. Source package directory where files are located.}

\item{dir_out}{Character. Target site directory where files are written.}
}
\value{
None. Side-effect of writing several files to disk.
}
\description{
Build content files such reference or vignette, converting from
Markdown to Quarto
}
