% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_reference_index}
\alias{build_reference_index}
\title{Build reference index from files in source package help folder}
\usage{
build_reference_index(dir_in, exclude = NULL, dir_out)
}
\arguments{
\item{dir_in}{Character. Help/reference file folder of source package.}

\item{exclude}{Character. Files to exclude from building process.
Specified as a regular expression.}

\item{dir_out}{Character. Reference foler of target documentation site.}
}
\value{
None. Side-effect of writing a file to disk
}
\description{
Build reference index from files in source package help folder
}
