% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_site}
\alias{build_site}
\title{Build Quarto site from Stata package files}
\usage{
build_site(pkg_dir, site_dir, rm_old_site_dir = TRUE)
}
\arguments{
\item{pkg_dir}{Charcter. Directory of the source package.}

\item{site_dir}{Charcter. Directory of the target site.}

\item{rm_old_site_dir}{Boolean. If \code{TRUE}, delete old site. Otherwise, keep.}
}
\description{
This single function performs several actions:
\itemize{
\item Creates folders for the Quarto site
\item Converts Markdown files to Quarto format, using some opinionated methods
\item Builds a command reference index, creating a table of functions, links,
and descriptions
\item Copies the package logo over
\item Composes and writes a Quarto YAML file, implementing an opinionated
layout and feeding forward some package and assets into relevant
places
\item Opens a preview of the Quarto documentation website
}
}
