% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{check_has_yaml}
\alias{check_has_yaml}
\title{Check whether a file has YAML frontmatter}
\usage{
check_has_yaml(file)
}
\arguments{
\item{file}{Character. Path to the markdown file}
}
\description{
Tries to find a pair of \verb{---} in the file
First one must be on the first line
}
