% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{check_title_in_yaml}
\alias{check_title_in_yaml}
\title{Check whether title is part of YAML}
\usage{
check_title_in_yaml(file)
}
\arguments{
\item{file}{Character. Path to the markdown file}
}
\description{
Check whether title is part of YAML
}
