% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_md_to_qmd}
\alias{convert_md_to_qmd}
\title{Convert Markdown files to Quarto}
\usage{
convert_md_to_qmd(path_in, dir_out)
}
\arguments{
\item{path_in}{Character. Path to Markdown file}

\item{dir_out}{Character. Directory to save Quarto file}
}
\description{
Convert Markdown files to Quarto
}
