% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_readme_to_index}
\alias{convert_readme_to_index}
\title{Convert package README to site index.qmd}
\usage{
convert_readme_to_index(pkg_dir, site_dir)
}
\arguments{
\item{pkg_dir}{Charcter. Directory of the source package.}

\item{site_dir}{Charcter. Directory of the target site.}
}
\description{
Simply copy the README file in the package directory to
the index file in the site directory, overwriting if it already exists.
}
