% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{create_folders}
\alias{create_folders}
\title{Create Quarto site folders}
\usage{
create_folders(site_dir, rm_old_site_dir = FALSE)
}
\arguments{
\item{site_dir}{Charcter. Directory of the target site.}

\item{rm_old_site_dir}{Boolean. If \code{TRUE}, delete old site. Otherwise, keep.}
}
\description{
Create folders for site, reference, articles, and images (e.g., logos)
}
