% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{extract_cmd_name_from_md}
\alias{extract_cmd_name_from_md}
\title{Extract Stata command name from \code{md} help file}
\usage{
extract_cmd_name_from_md(file_lines)
}
\arguments{
\item{file_lines}{Character vector of lines read from a file.}
}
\value{
Character. If command name found, that text. Otherwise, empty
string.
}
\description{
Based on the assumption that the \code{md} help file is generated by adodown
or follows its format, the function extracts the command name where it is
expected to be found.
}
