% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_custom_table.R
\name{find_titles}
\alias{find_titles}
\title{Find the index of the "title" key in YAML}
\usage{
find_titles(yaml)
}
\arguments{
\item{yaml}{YAML file loaded by \code{yaml::yaml.load_file()}}
}
\description{
Find the index of the "title" key in YAML
}
