% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{get_cmd_short_desc}
\alias{get_cmd_short_desc}
\title{Get short description of command from help file}
\usage{
get_cmd_short_desc(file)
}
\arguments{
\item{file}{Character. Path to source help file}
}
\value{
Character. Short description of command from help file title
}
\description{
The command short description has the following format:
\verb{commandname - This command is used for short description.}
}
