% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{make_article_yaml}
\alias{make_article_yaml}
\title{Make article component of Quarto YAML}
\usage{
make_article_yaml(articles_dir)
}
\arguments{
\item{articles_dir}{Character. Directory where package articles are stored}
}
\description{
Performs the following steps:
\itemize{
\item Compiles articles in the source package
\item Composes the nested list structure expected by \code{{yaml::as.yaml}}
}
}
