% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reference_index.R
\name{write_contents}
\alias{write_contents}
\title{Write the table contents of a section/sub-section of documentation}
\usage{
write_contents(dir_in, yaml, index)
}
\arguments{
\item{dir_in}{Character. Path to root of Stata package directory.}

\item{yaml}{YAML file loaded by \code{yaml::yaml.load_file()}}

\item{index}{Index in the YAML of the object to write.}
}
\description{
Composes the table of commands, links to their docs, and description
of their use
}
\seealso{
build_reference_index
}
