% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{write_quarto_yaml}
\alias{write_quarto_yaml}
\title{Write Quarto YAML to disk}
\usage{
write_quarto_yaml(yaml, site_dir)
}
\arguments{
\item{yaml}{Character. YAML specification string}

\item{site_dir}{Character. Directory where \verb{_quarto.yml} should be written}
}
\description{
Write Quarto YAML to disk
}
