% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reference_index.R
\name{write_sub_section}
\alias{write_sub_section}
\title{Write a sub-section of documentation}
\usage{
write_sub_section(dir_in, yaml, index)
}
\arguments{
\item{dir_in}{Character. Path to root of Stata package directory.}

\item{yaml}{YAML file loaded by \code{yaml::yaml.load_file()}}

\item{index}{Index in the YAML of the object to write.}
}
\description{
Sub-sections contain one or more of the following:
\itemize{
\item Subtitle (required)
\item Description (optional)
\item Contents (required)
}
}
