% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{compose_quarto_yaml}
\alias{compose_quarto_yaml}
\title{Compose YAML for Quarto package website}
\usage{
compose_quarto_yaml(pkg_dir)
}
\arguments{
\item{pkg_dir}{Character. Path to root of target path}
}
\value{
Character. Quarto YAML
}
\description{
Compose YAML for Quarto package website
}
